+++ docs/src/changelog.md
@@ -2,6 +2,7 @@
 
 ## Version 0.4.11 (2024-05-01)
 ### New Features
+- make changes to imported Matpower networks after import.
 
 ### Enhancements
 - added functions to facilitate the creation of networks
@@ -24,3 +25,12 @@
 ### Bug Fixes
 - calculation of branch flow and losses
 - branches out of service
+
+## Version 0.4.8 (2024-03-26)
+- first package release registered in the Julia registry
+
+## Version 0.4.1 (2023-12-19)
+- Initial release of Sparlectra
+
+## Version 0.4.0 (2023-11-30)
+- Initial public commit of Sparlectra